--- Cloud Computing/connect appengine-cloudsql/app.standard.yaml
@@ -1,26 +1,0 @@
-# Copyright 2018 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-runtime: python37
-entrypoint: gunicorn -b :$PORT app:app
-
-# Note: Saving credentials in environment variables is convenient, but not
-# secure - consider a more secure solution such as
-# Cloud Secret Manager (https://cloud.google.com/secret-manager) to help
-# keep secrets safe.
-env_variables:
-  INSTANCE_UNIX_SOCKET: /cloudsql/project-bangkit-352403:asia-southeast2:bangkit-tez
-  DB_USER: master
-  DB_PASS: 123
-  DB_NAME: demo
