+++ Cloud Computing/connect appengine-cloudsql/app.yaml
@@ -1,0 +1,11 @@
+# [START gae_python37_cloudsql_config]
+runtime: python37
+
+instance_class : F4_1G
+
+env_variables:
+  CLOUD_SQL_USERNAME: master
+  CLOUD_SQL_PASSWORD: 123
+  CLOUD_SQL_DATABASE_NAME: demo
+  CLOUD_SQL_CONNECTION_NAME: 'project-bangkit-352403:asia-southeast2:bangkit-tez'
+# [END gae_python37_cloudsql_config]
