+++ Cloud Computing/connect appengine-cloudsql/main.py
@@ -1,67 +1,122 @@
-import datetime
-import logging
+# [START gae_python37_cloudsql_mysql]
 import os
-from typing import Dict
 
-from flask import Flask, render_template, request, Response
+from flask import Flask,jsonify,request
+import pymysql
 
-import sqlalchemy
-
-from connect_connector import connect_with_connector
-from connect_tcp import connect_tcp_socket
-from connect_unix import connect_unix_socket
+db_user = os.environ.get('CLOUD_SQL_USERNAME')
+db_password = os.environ.get('CLOUD_SQL_PASSWORD')
+db_name = os.environ.get('CLOUD_SQL_DATABASE_NAME')
+db_connection_name = os.environ.get('CLOUD_SQL_CONNECTION_NAME')
 
 app = Flask(__name__)
 
-logger = logging.getLogger()
+#test server
+@app.route('/')
+def hello():
+    return 'Hello World!'
 
-def init_connection_pool() -> sqlalchemy.engine.base.Engine:
-    # use a TCP socket when INSTANCE_HOST (e.g. 127.0.0.1) is defined
-    if os.environ.get("INSTANCE_HOST"):
-        return connect_tcp_socket()
-
-    # use a Unix socket when INSTANCE_UNIX_SOCKET (e.g. /cloudsql/project:region:instance) is defined
-    if os.environ.get("INSTANCE_UNIX_SOCKET"):
-        return connect_unix_socket()
-
-    # use the connector when INSTANCE_CONNECTION_NAME (e.g. project:region:instance) is defined
-    if os.environ.get("INSTANCE_CONNECTION_NAME"):
-        return connect_with_connector()
-
-    raise ValueError(
-        "Missing database connection type. Please define one of INSTANCE_HOST, INSTANCE_UNIX_SOCKET, or INSTANCE_CONNECTION_NAME"
-    )
-db = None
-
-@app.before_first_request
-def init_db() -> sqlalchemy.engine.base.Engine:
-    global db
-    db = init_connection_pool()
-    migrate_db(db)
+# route login
+@app.route('/auth/login')
+def login():
+    username = str(request.args.get('username'))
+    password = str(request.args.get('password'))
+    #connect to cloud sql
+    if os.environ.get('GAE_ENV') == 'standard':
+        # If deployed, use the local socket interface for accessing Cloud SQL
+        unix_socket = '/cloudsql/{}'.format(db_connection_name)
+        cnx = pymysql.connect(user=db_user, password=db_password,
+                              unix_socket=unix_socket, db=db_name)
+    else:
+        host = '127.0.0.1'
+        cnx = pymysql.connect(user=db_user, password=db_password,
+                              host=host, db=db_name)
+    with cnx.cursor() as cursor:
+        cursor.execute('SELECT * FROM USER WHERE username = %s', username)
+        result = cursor.fetchone()
+        if result is not None:
+            if result[1] == password:
+                js={
+                    "code":"Login sukses";
+                }
+            else:
+                js={
+                    code:"Login gagal";
+                }
+        else:
+            js={
+                code:"Login gagal";
+            }
+        return jsonify(js)
+    cnx.close()
     
-def get_index_context(db: sqlalchemy.engine.base.Engine) -> Dict:
-    demo_txt = []
+# route register
+@app.route('/auth/register')
+def register():
+    username = str(request.args.get('username'))
+    password = str(request.args.get('password'))
+    #connect to cloud sql
+    if os.environ.get('GAE_ENV') == 'standard':
+        # If deployed, use the local socket interface for accessing Cloud SQL
+        unix_socket = '/cloudsql/{}'.format(db_connection_name)
+        cnx = pymysql.connect(user=db_user, password=db_password,
+                              unix_socket=unix_socket, db=db_name)
+    else:
+        host = '127.0.0.1'
+        cnx = pymysql.connect(user=db_user, password=db_password,
+                              host=host, db=db_name)
+    with cnx.cursor() as cursor:
+        cursor.execute("INSERT INTO USER (username, password) VALUES (%s, %s)", (username, password))
+        result = cursor.fetchone()
+        cnx.commit()
+    cnx.close()
+    with cnx.cursor() as cursor:
+        cursor.execute('SELECT * FROM USER WHERE username = %s and password =%s', username, password)
+        result = cursor.fetchone()
+        if result is not None:
+            if result[1] == password:
+                js={
+                    "code":"Login sukses";
+                }
+            else:
+                js={
+                    code:"Login gagal";
+                }
+        else:
+            js={
+                code:"Login gagal";
+            }
+        return jsonify(js)
+    cnx.close()
+            
+@app.route('/db')
+def main():
+    # When deployed to App Engine, the `GAE_ENV` environment variable will be
+    # set to `standard`
+    if os.environ.get('GAE_ENV') == 'standard':
+        # If deployed, use the local socket interface for accessing Cloud SQL
+        unix_socket = '/cloudsql/{}'.format(db_connection_name)
+        cnx = pymysql.connect(user=db_user, password=db_password,
+                              unix_socket=unix_socket, db=db_name)
+    else:
+        # If running locally, use the TCP connections instead
+        # Set up Cloud SQL Proxy (cloud.google.com/sql/docs/mysql/sql-proxy)
+        # so that your application can use 127.0.0.1:3306 to connect to your
+        # Cloud SQL instance
+        host = '127.0.0.1'
+        cnx = pymysql.connect(user=db_user, password=db_password,
+                              host=host, db=db_name)
 
-    with db.connect() as conn:
-        # Execute the query and fetch all results
-        recent_demo = conn.execute(
-            "SELECT demo_txt FROM demo_tbl"
-        ).fetchall()
-        # Convert the results into a list of dicts representing votes
-        for row in recent_demo:
-            demo.append({"demo_txt": row[0]})
+    with cnx.cursor() as cursor:
+        #query pending
+        cursor.execute('SELECT demo_db FROM demo_tbl;)
+        result = cursor.fetchall()
+        current_msg = result[0][0]
+    cnx.close()
 
-                # stmt = sqlalchemy.text(
-                #     "SELECT COUNT(demo_id) FROM demo WHERE demo_txt=:demo_txt"
-                # )
-                # # Count number of votes for tabs
-                # tab_result = conn.execute(stmt, candidate="TABS").fetchone()
-                # tab_count = tab_result[0]
-                # # Count number of votes for spaces
-                # space_result = conn.execute(stmt, candidate="SPACES").fetchone()
-                # space_count = space_result[0]
+    return str(current_msg)
+# [END gae_python37_cloudsql_mysql]
 
-    return (recent_demo)
-    
-if __name__ == "__main__":
-    app.run(host="127.0.0.1", port=8080, debug=True)
+
+if __name__ == '__main__':
+    app.run(host='127.0.0.1', port=8080, debug=True)
