--- Cloud Computing/connect appengine-cloudsql/service.yaml
@@ -1,30 +1,0 @@
-# Copyright 2021 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# [START cloud_sql_mysql_sqlalchemy_gke_quickstart_service]
-# The service provides a load-balancing proxy over the gke-cloud-sql-app
-# pods. By specifying the type as a 'LoadBalancer', Kubernetes Engine will
-# create an external HTTP load balancer.
-apiVersion: v1
-kind: Service
-metadata:
-  name: gke-cloud-sql-app
-spec:
-  type: LoadBalancer
-  selector:
-    app: gke-cloud-sql-app
-  ports:
-  - port: 80
-    targetPort: 8080
-# [END cloud_sql_mysql_sqlalchemy_gke_quickstart_service]
